requirements.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
